cerbero: Add support for building Meson projects with MSVC
Select one of the two custom configs (win32-msvc, win64-msvc) which will enable
the 'visualstudio' variant, and if the build system being used is Meson, an MSVC
environment will be set using Visual Studio's vcvarsall.bat. This will cause
Meson to auto-detect the Visual Studio toolchain and use it.

Cerbero sets C_INCLUDE_PATH and LIBRARY_PATH so that gcc/ld can find headers and
libraries when invoked for configure. Hence, when building with MSVC, we also
append to the corresponding INCLUDE and LIB variables.

This is done in `modify_environment` instead of the config files because it's
parsed before `prefix` is set.