cerbero: Fix shell.which on Windows and use it to find Meson
On Windows, an executable name is never the same as the basename of its
path. Executables are detected by looking at the list of extensions in the
%PATHEXT% environment variable.

For compatibility with how MinGW bash works, we also first look for a file
with the requested name in %PATH%.

With this, we can now find the meson.py Python script and use it.