Replace gettext.recipe with proxy-libintl.recipe
We only use gettext on non-Linux systems on which gettext takes forever
to build and produces a libintl that no one actually uses for doing
l10n/i18n because the setup is not easy to do even if you know how the
l10n stack works.

proxy-libintl can work in two modes: stubs-only and library-loader. In
stub-only mode (-DSTUBS_ONLY -- the default in the recipe), gettext
functions do nothing. In library-loader mode (without -DSTUBS_ONLY), it
looks for a libintl shared library to load dynamically at runtime which
is supposed to provide the actual implementation.

Also, port various recipes to use this new proxy-libintl.

TODO: Ideally, we want to statically link this into glib and look for a
libintl library and if it's not found, insert stubs.