cerbero: Add a new option to fetch: --print-only
This will print all source URLs and tarball names to stdout with
a prefix denoting their type. For now, only prints tarballs.

The tarball URLs have all been URL-encoded so that, for instance, spaces
are converted to %20. This makes it safer to parse stdout for the URL
and tarball name. Note that the tarball name can still contain spaces.

The recommended way of parsing the output is:

* Check if it begins with 'TARBALL: '
* If it does, read till the next space. This is your URL.
* Skip the space. Everything till newline is your tarball name.

This will be used to implement our tarball mirror setup on fdo.